+++ src/data/consume/gecko-potion.ts
@@ -5,23 +5,12 @@ import { AssistantMessage } from "message.ts";
 export const actions: AssistantAction[] = [
     {
         trigger: "consume",
-        predicate: [
-            {
-                or: [
-                    "consumable:elixir-of-life-minor",
-                    "consumable:elixir-of-life-lesser",
-                    "consumable:elixir-of-life-moderate",
-                    "consumable:elixir-of-life-greater",
-                    "consumable:elixir-of-life-major",
-                    "consumable:elixir-of-life-true",
-                ],
-            },
-        ],
+        predicate: ["consumable:gecko-potion"],
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             const target = message.target?.actor ?? message.speaker.actor;
 
-            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.lPRuIRbu0rHBkoKY", {
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.buJnkFBzL4e22ASp", {
                 _id: null,
                 system: {
                     context: {