+++ src/data/consume/insight-coffee.ts
@@ -1,0 +1,81 @@
+import { AssistantAction } from "action.ts";
+import { EffectSource } from "foundry-pf2e";
+import { AssistantMessage } from "message.ts";
+
+export const actions: AssistantAction[] = [
+    {
+        trigger: "consume",
+        predicate: ["consumable:insight-coffee-lesser"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.XGHrqFn9QXdTM42R", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:insight-coffee-moderate"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.kkXy5JWC7OTBqTKg", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:insight-coffee-greater"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.q6Qq3Ie3M22XI5bw", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+];