+++ src/data/consume/sea-touch-elixir.ts
@@ -1,0 +1,81 @@
+import { AssistantAction } from "action.ts";
+import { EffectSource } from "foundry-pf2e";
+import { AssistantMessage } from "message.ts";
+
+export const actions: AssistantAction[] = [
+    {
+        trigger: "consume",
+        predicate: ["consumable:sea-touch-elixir-lesser"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.thOpQunbQr77XWdF", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:sea-touch-elixir-moderate"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.9keegq0GdS1eSrNr", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:sea-touch-elixir-greater"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.RRusoN3HEGnDO1Dg", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+];