+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [6.1.2] - 2025-01-11
+
+### Added
+
+- Assassin Vine Wine
+- Blood Booster
+- Bomber's Eye Elixir
+- Bravo's Brew
+- Chromatic Jellyfish Oil
+- Darkvision Elixir
+- Egg Cream Fizz
+- Fury Cocktail
+- Gecko Potion
+- Insight Coffee
+- Merciful Balm
+- Potency Crystal
+- Potion of Emergency Escape
+- Predator's Claw
+- Sea Touch Elixir
+- Shielding Salve
+- Shining Ammunition
+- Soothing Toddy
+- Soothing Tonic
+- Spiderfoot Brew
+- Stone Body Mutagen
+- Stone Fist Elixir
+- Viperous Elixir
+
 ## [6.1.1] - 2025-01-11
 
 ### Added
@@ -92,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Auto Self-Applied Effects (This is disabled if PF2e Toolbelt is doing this as well)
 - Swashbuckler Panache
 
-[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.1.1...HEAD
+[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.1.2...HEAD
+[6.1.2]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.1.1...v6.1.2
 [6.1.1]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.1.0...v6.1.1
 [6.1.0]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.0.4...v6.1.0
 [6.0.4]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v6.0.3...v6.0.4